+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2023-12-07
+
+### Updated
+
+- Library upgrades to address security vulnerabilities.
+- Fix an issue with sample workflow configurations where `textract` workflow was missing before any `entity` based detection workflow.
+- Upgrade AWS Lambda runtimes to Python 3.12, Nodejs 20, and Java 21.
+- Update AWS SDK and AWS CDK versions.
+
 ## [1.0.2] - 2023-11-09
 
 ### Updated